@@ -17,7 +17,7 @@
 package com.yjh.rer.util;
 
 import android.os.SystemClock;
-import android.support.v4.util.ArrayMap;
+import android.util.ArrayMap;
 
 import java.util.concurrent.TimeUnit;
 
